log170524

+++ real-coffee-backend/database/migrations/2024_05_01_082332_create_carts_table.php
@@ -13,9 +13,9 @@ return new class extends Migration
     {
         Schema::create('carts', function (Blueprint $table) {
             $table->id();
-            $table->integer('phonenumber');
+            $table->bigInteger('phonenumber');
             $table->integer('product_id');
-            // $table->integer('product_qty');
+            // $table->integer('producttype');
             $table->timestamps();
         });
     }
